Tests: Clean up memory leak from nock
We were setting up nock's disableNetConnect globally, but never cleaning it up. Use a afterAll hook to clean up the interceptors and restore the normal functionality and a beforeAll hook to activate nock.

The disableNetConnect call _cannot_ be inside the beforeAll hook due to some kind of race between superagent and nock. Moving it there allows some superagent requests to escape nock's sandbox and randomly break the test runner.